+++ MLEM/Input/Keybind.cs
@@ -246,42 +246,6 @@ namespace MLEM.Input {
         }
 
         /// <summary>
-        /// Returns the amount of time that this keybind has been held down for.
-        /// If this input isn't currently down, this method returns <see cref="TimeSpan.Zero"/>.
-        /// </summary>
-        /// <param name="handler">The input handler to query the keys with</param>
-        /// <param name="gamepadIndex">The index of the gamepad to query, or -1 to query all gamepads</param>
-        /// <returns>The resulting down time, or <see cref="TimeSpan.Zero"/> if the input is not being held.</returns>
-        [Obsolete("This method is deprecated, as it does not query Modifiers or InverseModifiers. Use InputHandler.GetDownTime or custom handling instead.")]
-        public TimeSpan GetDownTime(InputHandler handler, int gamepadIndex = -1) {
-            return this.Combinations.Max(c => c.GetDownTime(handler, gamepadIndex));
-        }
-
-        /// <summary>
-        /// Returns the amount of time that this keybind has been up for since the last time it was down.
-        /// If this input isn't currently up, this method returns <see cref="TimeSpan.Zero"/>.
-        /// </summary>
-        /// <param name="handler">The input handler to query the keys with</param>
-        /// <param name="gamepadIndex">The index of the gamepad to query, or -1 to query all gamepads</param>
-        /// <returns>The resulting up time, or <see cref="TimeSpan.Zero"/> if the input is being held.</returns>
-        [Obsolete("This method is deprecated, as it does not query Modifiers or InverseModifiers. Use InputHandler.GetUpTime or custom handling instead.")]
-        public TimeSpan GetUpTime(InputHandler handler, int gamepadIndex = -1) {
-            return this.Combinations.Min(c => c.GetUpTime(handler, gamepadIndex));
-        }
-
-        /// <summary>
-        /// Returns the amount of time that has passed since this keybind last counted as pressed.
-        /// If this input hasn't been pressed previously, or is currently pressed, this method returns <see cref="TimeSpan.Zero"/>.
-        /// </summary>
-        /// <param name="handler">The input handler to query the keys with</param>
-        /// <param name="gamepadIndex">The index of the gamepad to query, or -1 to query all gamepads</param>
-        /// <returns>The resulting up time, or <see cref="TimeSpan.Zero"/> if the input has never been pressed, or is currently pressed.</returns>
-        [Obsolete("This method is deprecated, as it does not query Modifiers or InverseModifiers. Use InputHandler.GetTimeSincePress or custom handling instead.")]
-        public TimeSpan GetTimeSincePress(InputHandler handler, int gamepadIndex = -1) {
-            return this.Combinations.Min(c => c.GetTimeSincePress(handler, gamepadIndex));
-        }
-
-        /// <summary>
         /// Returns an enumerable of all of the combinations that this keybind currently contains
         /// </summary>
         /// <returns>This keybind's combinations</returns>
@@ -502,41 +466,6 @@ namespace MLEM.Input {
                 return true;
             }
 
-            /// <summary>
-            /// Returns the amount of time that this combination has been held down for.
-            /// If this input isn't currently down, this method returns <see cref="TimeSpan.Zero"/>.
-            /// </summary>
-            /// <param name="handler">The input handler to query the keys with</param>
-            /// <param name="gamepadIndex">The index of the gamepad to query, or -1 to query all gamepads</param>
-            /// <returns>The resulting down time, or <see cref="TimeSpan.Zero"/> if the input is not being held.</returns>
-            [Obsolete("This method is deprecated, as it does not query Modifiers or InverseModifiers. Use InputHandler.GetDownTime or custom handling instead.")]
-            public TimeSpan GetDownTime(InputHandler handler, int gamepadIndex = -1) {
-                return handler.GetDownTime(this.Key, gamepadIndex);
-            }
-
-            /// <summary>
-            /// Returns the amount of time that this combination has been up for since the last time it was down.
-            /// If this input isn't currently up, this method returns <see cref="TimeSpan.Zero"/>.
-            /// </summary>
-            /// <param name="handler">The input handler to query the keys with</param>
-            /// <param name="gamepadIndex">The index of the gamepad to query, or -1 to query all gamepads</param>
-            /// <returns>The resulting up time, or <see cref="TimeSpan.Zero"/> if the input is being held.</returns>
-            [Obsolete("This method is deprecated, as it does not query Modifiers or InverseModifiers. Use InputHandler.GetUpTime or custom handling instead.")]
-            public TimeSpan GetUpTime(InputHandler handler, int gamepadIndex = -1) {
-                return handler.GetUpTime(this.Key, gamepadIndex);
-            }
-
-            /// <summary>
-            /// Returns the amount of time that has passed since this combination last counted as pressed.
-            /// If this input hasn't been pressed previously, or is currently pressed, this method returns <see cref="TimeSpan.Zero"/>.
-            /// </summary>
-            /// <param name="handler">The input handler to query the keys with</param>
-            /// <param name="gamepadIndex">The index of the gamepad to query, or -1 to query all gamepads</param>
-            /// <returns>The resulting up time, or <see cref="TimeSpan.Zero"/> if the input has never been pressed, or is currently pressed.</returns>
-            [Obsolete("This method is deprecated, as it does not query Modifiers or InverseModifiers. Use InputHandler.GetTimeSincePress or custom handling instead.")]
-            public TimeSpan GetTimeSincePress(InputHandler handler, int gamepadIndex = -1) {
-                return handler.GetTimeSincePress(this.Key, gamepadIndex);
-            }
 
             /// <summary>
             /// Converts this combination into an easily human-readable string.